implement: bug-squash-cursor-pagination-drops-rows
Change ListCommentsForSession and ListSessionsForOrgWithCursor to use a
true (created_at, id) keyset instead of a strict created_at bound.
SQLite uses anonymous ? with the boundary created_at passed twice;
Postgres reuses $N. Add LastID to domain params structs; thread it
through both adapters. First-page path uses a future time + max-id
sentinel so all page-1 rows qualify. Regenerated sqlc for both dialects.

+++ internal/db/pgstore/querier.go
@@ -14,8 +14,8 @@ import (
 type Querier interface {
 	AddOrgMember(ctx context.Context, arg AddOrgMemberParams) error
 	AddSessionMember(ctx context.Context, arg AddSessionMemberParams) error
-	AllocateNextSeq(ctx context.Context, sessionID string) (int32, error)
-	AllocateNextSeqN(ctx context.Context, arg AllocateNextSeqNParams) (int32, error)
+	AllocateNextSeq(ctx context.Context, sessionID string) (int64, error)
+	AllocateNextSeqN(ctx context.Context, arg AllocateNextSeqNParams) (int64, error)
 	CleanupExpiredOAuthState(ctx context.Context, expiresAt time.Time) error
 	ClearFinalizeLock(ctx context.Context, arg ClearFinalizeLockParams) error
 	ConsumeMagicLinkToken(ctx context.Context, arg ConsumeMagicLinkTokenParams) error
